chore($package): relocate package in config
- integrate configs components
