+++ module-core/src/main/java/com/woomoolmarket/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.woomoolmarket.config.auditing;
+package com.woomoolmarket.config;
 
 import java.util.Optional;
 import org.springframework.context.annotation.Bean;
